hyper_dmabuf: fix map failure issue when assign 4G memory to UOS (v2)
When assign 4G memory to UOS, 32bit GPA ref_handle will overflow, instead
we need set ref_handle as unsigned long to fix this issue. op is int type,
then we need use two int op[7] and op[8] to pass ref_handle.

v2 changes:
- Aligned Xen backend with backend interface changes

+++ drivers/dma-buf/hyper_dmabuf/hyper_dmabuf_drv.h
@@ -83,7 +83,7 @@ struct hyper_dmabuf_bknd_ops {
 	int (*get_vm_id)(void);
 
 	/* get pages shared via hypervisor-specific method */
-	int (*share_pages)(struct page **, int, int, void **);
+	long (*share_pages)(struct page **, int, int, void **);
 
 	/* make shared pages unshared via hypervisor specific method */
 	int (*unshare_pages)(void **, int);
